fix(contracts): align day0 contracts with M3 spec (55-24913)

+++ contracts/src/main/java/com/team28/booking/contracts/events/BookingCompletedEvent.java
@@ -1,6 +1,4 @@
 package com.team28.booking.contracts.events;
 
-import java.math.BigDecimal;
-
-public record BookingCompletedEvent(Long bookingId, Long userId, Long providerId, BigDecimal totalPrice) {
+public record BookingCompletedEvent(Long bookingId, Long userId, Long providerId, Double totalPrice) {
 }
