+++ contracts/src/main/java/com/team28/booking/contracts/events/PaymentCompletedEvent.java
@@ -1,6 +1,4 @@
 package com.team28.booking.contracts.events;
 
-import java.math.BigDecimal;
-
-public record PaymentCompletedEvent(Long invoiceId, Long bookingId, BigDecimal amount) {
+public record PaymentCompletedEvent(Long invoiceId, Long bookingId, Double amount) {
 }
