+++ contracts/src/main/java/com/team28/booking/contracts/events/PaymentInitiatedEvent.java
@@ -1,6 +1,4 @@
 package com.team28.booking.contracts.events;
 
-import java.math.BigDecimal;
-
-public record PaymentInitiatedEvent(Long invoiceId, Long bookingId, BigDecimal amount) {
+public record PaymentInitiatedEvent(Long invoiceId, Long bookingId, Double amount) {
 }
