+++ contracts/src/main/java/com/team28/booking/contracts/events/PaymentRefundedEvent.java
@@ -1,6 +1,4 @@
 package com.team28.booking.contracts.events;
 
-import java.math.BigDecimal;
-
-public record PaymentRefundedEvent(Long invoiceId, Long bookingId, BigDecimal refundAmount) {
+public record PaymentRefundedEvent(Long invoiceId, Long bookingId, Double refundAmount) {
 }
